refactor: useClickOutside 훅 추출
- Header, ProfileImagePopover, Dialog 3곳의 중복 mousedown click-outside 패턴을 공용 훅으로 통합
- disabled, excludeRefs 옵션으로 조건부 활성화 및 트리거 요소 제외 지원
